+++ hutool-core/src/main/java/cn/hutool/core/collection/CollUtil.java
@@ -1299,14 +1299,14 @@ public class CollUtil {
 	 * @param <E>              集合元素类型
 	 * @param resultCollection 存放移除结果的集合
 	 * @param targetCollection 被操作移除元素的集合
-	 * @param filter           用于是否移除判断的过滤器
+	 * @param predicate           用于是否移除判断的过滤器
 	 */
-	public static <T extends Collection<E>, E> T removeWithAddIf(T targetCollection, T resultCollection, Predicate<? super E> filter) {
-		Objects.requireNonNull(filter);
+	public static <T extends Collection<E>, E> T removeWithAddIf(T targetCollection, T resultCollection, Predicate<? super E> predicate) {
+		Objects.requireNonNull(predicate);
 		final Iterator<E> each = targetCollection.iterator();
 		while (each.hasNext()) {
 			E next = each.next();
-			if (filter.test(next)) {
+			if (predicate.test(next)) {
 				resultCollection.add(next);
 				each.remove();
 			}
@@ -1315,20 +1315,20 @@ public class CollUtil {
 	}
 
 	/**
-	 * 移除集合中的多个元素，并将结果存放到生成的新集合中后返回
+	 * 移除集合中的多个元素，并将结果存放到生成的新集合中后返回<br>
 	 * 此方法直接修改原集合
 	 *
 	 * @param <T>              集合类型
 	 * @param <E>              集合元素类型
 	 * @param targetCollection 被操作移除元素的集合
-	 * @param filter           用于是否移除判断的过滤器
+	 * @param predicate        用于是否移除判断的过滤器
 	 * @return 移除结果的集合
+	 * @since 5.7.17
 	 */
-	@SuppressWarnings("unchecked")
-	public static <T extends Collection<E>, E> T removeWithAddIf(T targetCollection, Predicate<? super E> filter) {
-		Collection<E> resultCollection = new ArrayList<>();
-		removeWithAddIf(targetCollection, resultCollection, filter);
-		return (T) resultCollection;
+	public static <T extends Collection<E>, E> List<E> removeWithAddIf(T targetCollection, Predicate<? super E> predicate) {
+		final List<E> removed = new ArrayList<>();
+		removeWithAddIf(targetCollection, removed, predicate);
+		return removed;
 	}
 
 	/**