+++ hutool-core/src/test/java/cn/hutool/core/util/NumberUtilTest.java
@@ -412,4 +412,26 @@ public class NumberUtilTest {
 		final Integer[] integers = NumberUtil.generateBySet(10, 100, 5);
 		Assert.assertEquals(5, integers.length);
 	}
+
+	@Test
+	public void isOddOrEvenTest(){
+		int[] a = { 0, 32, -32, 123, -123 };
+		Assert.assertFalse(NumberUtil.isOdd(a[0]));
+		Assert.assertTrue(NumberUtil.isEven(a[0]));
+
+		Assert.assertFalse(NumberUtil.isOdd(a[1]));
+		Assert.assertTrue(NumberUtil.isEven(a[1]));
+
+		Assert.assertFalse(NumberUtil.isOdd(a[2]));
+		Assert.assertTrue(NumberUtil.isEven(a[2]));
+
+		Assert.assertTrue(NumberUtil.isOdd(a[3]));
+		Assert.assertFalse(NumberUtil.isEven(a[3]));
+
+		Assert.assertTrue(NumberUtil.isOdd(a[4]));
+		Assert.assertFalse(NumberUtil.isEven(a[4]));
+	}
+
+
+
 }